+++ packages/library/src/commonMain/kotlin/io/realm/BaseRealm.kt
@@ -14,8 +14,6 @@
  */
 package io.realm
 
-import kotlin.reflect.KClass
-
 /**
  * Base class for all Realm instances ([Realm] and [MutableRealm]).
  */
@@ -32,18 +30,6 @@ interface BaseRealm {
     var version: VersionId
 
     /**
-     * Returns the results of querying for all objects of a specific type.
-     *
-     * For a [Realm] instance this reflects the state of the realm at the invocation time, thus
-     * the results will not change on updates to the Realm. For a [MutableRealm] the result is live
-     * and will in fact reflect updates to the [MutableRealm].
-     *
-     * @param clazz The class of the objects to query for.
-     * @return The result of the query as of the time of invoking this method.
-     */
-    open fun <T : RealmObject> objects(clazz: KClass<T>): RealmResults<T>
-
-    /**
      * Returns the current number of active versions in the Realm file. A large number of active versions can have
      * a negative impact on the Realm file size on disk.
      *
@@ -58,8 +44,4 @@ interface BaseRealm {
      * @return `true` if the Realm has been closed. `false` if not.
      */
     fun isClosed(): Boolean
-}
-
-inline fun <reified T : RealmObject> BaseRealm.objects(): RealmResults<T> {
-    return this.objects(T::class)
 }